keystone(validators): decode canonical bytes; 0-based enum positions
Stage 4 of the Keystone cutover. Layer-C validators in
src/schemaValidators/ now decode the canonical ABI shape produced by
the generic encoder. 7 validators updated (the 6 that change bytes
plus arbitration-kleros for type-canonicalisation).

The 6 byte-changing validators — accept 0-based enum positions:
- FigaroGeoV2Validator (classOfService 0..3); also widens
  massGrams/volumeMl decode from uint32 to uint256 per the canonical
  rule (integer → uint256) with an explicit range check.
- FigaroFulfilmentV2Validator (modalities 0..3, coordinations 0..2,
  handoffPoints 0..3); MODALITY_DELIVERY 3→2.
- FigaroProximityPolicyV1Validator (bands 0..2)
- FigaroProximityProofV1Validator (band 0..2)
- FigaroOffsetPolicyV1Validator (providers 0..3)
- FigaroConsentV1Validator — structural change: documents decoded as
  ConsentDocument[] (tuple[]) per the canonical object-array rule,
  not three parallel struct-of-arrays. The ArrayLengthMismatch error
  is deleted (the tuple-of-arrays representation can't desync).

FigaroArbitrationKlerosV1Validator: bytes unchanged for real
subcourts (the prior commit's "none" insertion preserves the legacy
1..4 indices), but klerosMinJurors decode widens uint8→uint256 to
match the canonical integer encoding.

5 GHG validators (protocol/iso/pas/en/custom): scope decode widens
uint8→uint256, same canonical integer rule. Behaviour unchanged for
valid inputs (scope 0..3 fits both); the decode type is now
type-correct rather than relying on uint8's high-bytes-must-be-zero
check.

Foundry test fixtures regenerated for the 6 byte-changing schemas:
~140 test assertions updated to 0-based enum values; "zero index is
sentinel"/"out-of-range-above-max" tests realigned. Consent test
rewrites helpers from parallel-array builders to ConsentDocument[]
builders. Verify: forge test --via-ir → 483 pass / 0 fail / 1 skip
(the Polygon fork test).

+++ src/schemaValidators/FigaroGHGEN16258V1Validator.sol
@@ -10,7 +10,7 @@ import {ISchemaValidator} from "../ISchemaValidator.sol";
 ///         declaration of energy consumption and GHG emissions of transport
 ///         services).
 ///
-/// @dev Content ABI encoding: `abi.encode(uint8 scope)`.
+/// @dev Content ABI encoding: `abi.encode(uint256 scope)` (canonical: every `integer` field encodes as uint256).
 ///      scope: 0 = unset, 1 = direct (Scope 1), 2 = purchased energy (Scope 2),
 ///             3 = value chain (Scope 3)
 ///
@@ -21,10 +21,10 @@ import {ISchemaValidator} from "../ISchemaValidator.sol";
 contract FigaroGHGEN16258V1Validator is ISchemaValidator {
     bytes32 public constant override schemaId = keccak256("figaro-ghg-en-16258-v1");
 
-    uint8 internal constant MAX_SCOPE = 3;
+    uint256 internal constant MAX_SCOPE = 3;
 
     error SchemaIdMismatch(bytes32 got, bytes32 expected);
-    error InvalidScope(uint8 got);
+    error InvalidScope(uint256 got);
     /// @dev Runtime `content` must byte-equal the committed clause `sectionData`.
     error SectionDataMismatch();
 
@@ -41,7 +41,7 @@ contract FigaroGHGEN16258V1Validator is ISchemaValidator {
     {
         if (id != schemaId) revert SchemaIdMismatch(id, schemaId);
         if (keccak256(sectionData) != keccak256(content)) revert SectionDataMismatch();
-        uint8 scope = abi.decode(content, (uint8));
+        uint256 scope = abi.decode(content, (uint256));
         if (scope > MAX_SCOPE) revert InvalidScope(scope);
     }
 }
