+++ src/schemaValidators/FigaroGHGProtocolV1Validator.sol
@@ -9,7 +9,7 @@ import {ISchemaValidator} from "../ISchemaValidator.sol";
 ///         under the GHG Protocol Corporate Standard (and related WRI/WBCSD
 ///         GHG Protocol guidance documents).
 ///
-/// @dev Content ABI encoding: `abi.encode(uint8 scope)`.
+/// @dev Content ABI encoding: `abi.encode(uint256 scope)` (canonical: every `integer` field encodes as uint256).
 ///      scope: 0 = unset, 1 = direct (Scope 1), 2 = purchased energy (Scope 2),
 ///             3 = value chain (Scope 3)
 ///
@@ -22,10 +22,10 @@ import {ISchemaValidator} from "../ISchemaValidator.sol";
 contract FigaroGHGProtocolV1Validator is ISchemaValidator {
     bytes32 public constant override schemaId = keccak256("figaro-ghg-protocol-v1");
 
-    uint8 internal constant MAX_SCOPE = 3;
+    uint256 internal constant MAX_SCOPE = 3;
 
     error SchemaIdMismatch(bytes32 got, bytes32 expected);
-    error InvalidScope(uint8 got);
+    error InvalidScope(uint256 got);
     /// @dev Runtime `content` must byte-equal the committed clause `sectionData`.
     ///      A seller who signed a Scope 1 clause cannot later attest under a
     ///      different scope via this schema.
@@ -44,7 +44,7 @@ contract FigaroGHGProtocolV1Validator is ISchemaValidator {
     {
         if (id != schemaId) revert SchemaIdMismatch(id, schemaId);
         if (keccak256(sectionData) != keccak256(content)) revert SectionDataMismatch();
-        uint8 scope = abi.decode(content, (uint8));
+        uint256 scope = abi.decode(content, (uint256));
         if (scope > MAX_SCOPE) revert InvalidScope(scope);
     }
 }
